+++ core/src/main/java/shape/models/Shape.java
@@ -6,10 +6,10 @@ public class Shape {
     private String id;
     private Double latitude;
     private Double longitude;
-    private Integer sequence;
-    private Integer distanceTraveled;
+    private Long sequence;
+    private Long distanceTraveled;
     private Shape(){}
-    private Shape(String id, Double latitude, Double longitude, Integer sequence, Integer distanceTraveled) {
+    private Shape(String id, Double latitude, Double longitude, Long sequence, Long distanceTraveled) {
         this.id = id;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -17,7 +17,7 @@ public class Shape {
         this.distanceTraveled = distanceTraveled;
     }
 
-    public static Shape getInstance(String id, Double latitude, Double longitude, Integer sequence, Integer distanceTraveled) {
+    public static Shape getInstance(String id, Double latitude, Double longitude, Long sequence, Long distanceTraveled) {
         if (latitude < -90 || latitude > 90)
             throw new ShapeException("La latitud debe estar entre -90 y 90 grados.");
 
@@ -45,11 +45,11 @@ public class Shape {
         return longitude;
     }
 
-    public Integer getSequence() {
+    public Long getSequence() {
         return sequence;
     }
 
-    public Integer getDistanceTraveled() {
+    public Long getDistanceTraveled() {
         return distanceTraveled;
     }
 }
